generalize GPT3Classifier -> InContextClassifier
prepping for adding generic HF generative classifier
